now user have to login first and each user just can see and handle her/his own todos

+++ todo/views.py
@@ -5,18 +5,25 @@ from .forms import TodoForm
 
 
 def todo_list(request):
-    if request.method == 'GET':
-        todos = Todo.objects.all()
-        form = TodoForm()
 
-        context = {'todos': todos, 'form': form}
-        return render(request, 'todo/index.html', context=context)
-    
+    if request.method == 'GET':
+        if request.user.is_authenticated:
+            todos = Todo.objects.filter(owner=request.user)
+            form = TodoForm()
+            context = {'todos': todos, 'form': form}
+            return render(request, 'todo/index.html', context=context)
+        else:
+            return redirect('account:register')
+
     elif request.method == 'POST':
-        todos = Todo.objects.all()
         form = TodoForm(request.POST)
+        todos = Todo.objects.filter(owner=request.user)
+
         if form.is_valid():
-            form.save()
+            user = request.user
+            new_form = form.save(commit=False)
+            new_form.owner = user
+            new_form.save()
             form = TodoForm()
             context = {'todos': todos, 'form': form}
             return render(request, 'todo/index.html', context=context)
@@ -27,9 +34,21 @@ def todo_detail(request, pk):
     context = {'todo': todo}
 
     if request.method == 'POST':
-        todo.delete()
-        context = {'todo': todo}
-        return redirect('todo:todo-list')
+        if request.POST.get('delete') == 'Delete':
+            todo.delete()
+            context = {'todo': todo}
+            return redirect('todo:todo-list')
+        elif request.POST.get('in progress') == 'In Progress':
+            todo.status = 'I'
+            todo.save()
+        elif request.POST.get('pending') == 'Pending':
+            todo.status = 'P'
+            todo.save()
+        elif request.POST.get('done') == 'Done':
+            todo.status = 'D'
+            todo.save()
+        else:
+            pass
 
     return render(request, 'todo/detail.html', context=context)
